Add sbt-tpolecat plugin to turn on common compiler settings.
Especially fatal warnings!

+++ src/main/scala/ddb/debugger/z/MyControl.scala
@@ -24,8 +24,8 @@ case class MySliderControl()(implicit rt: DebuggerRuntime) extends ControlProvid
         case (_, idx) =>
           IO {
             runLater {
-              control.max = idx
-              control.value = idx
+              control.max = idx.toDouble
+              control.value = idx.toDouble
             }
           }
       })
